+++ service/tier1.go
@@ -23,6 +23,7 @@ import (
 	"github.com/streamingfast/dgrpc"
 	"github.com/streamingfast/dmetering"
 	"github.com/streamingfast/dmetrics"
+	"github.com/streamingfast/dsession"
 	"github.com/streamingfast/dstore"
 	"github.com/streamingfast/logging"
 	tracing "github.com/streamingfast/sf-tracing"
@@ -45,12 +46,10 @@ import (
 	"github.com/streamingfast/substreams/storage/execout"
 	"github.com/streamingfast/substreams/storage/store"
 	"github.com/streamingfast/substreams/wasm"
-	pbworker "github.com/streamingfast/worker-pool-protocol/pb/sf/worker/v1"
 	"go.opentelemetry.io/otel/attribute"
 	ttrace "go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -82,7 +81,7 @@ type Tier1Service struct {
 	activeRequestsSoftLimit int
 	activeRequestsHardLimit int
 	tier2RequestParameters  reqctx.Tier2RequestParameters
-	globalRequestPool       *GlobalRequestPool
+	sessionPool             dsession.SessionPool
 }
 
 func getBlockTypeFromStreamFactory(sf *StreamFactory) (string, error) {
@@ -141,17 +140,18 @@ func NewTier1(
 	appSetIsReadyState func(isReady bool),
 	substreamsClientConfig *client.SubstreamsClientConfig,
 	tier2RequestParameters reqctx.Tier2RequestParameters,
-	workerPoolFactory work.WorkerPoolFactory,
-
 	enforceCompression bool,
 	activeRequestsSoftLimit int,
 	activeRequestsHardLimit int,
 	sharedCacheSize uint64,
-	globalRequestPool *GlobalRequestPool,
+	sessionPool dsession.SessionPool,
 	opts ...Option,
 ) (*Tier1Service, error) {
 
 	clientFactory := client.NewInternalClientFactory(substreamsClientConfig)
+
+	// Create WorkerPoolFactory using the sessionPool
+	workerPoolFactory := work.NewSessionWorkerPoolFactory(sessionPool, clientFactory)
 
 	runtimeConfig := config.NewTier1RuntimeConfig(
 		stateBundleSize,
@@ -161,7 +161,7 @@ func NewTier1(
 		quickSaveStore,
 		defaultCacheTag,
 		clientFactory,
-		workerPoolFactory,
+		workerPoolFactory.WorkerPool,
 	)
 
 	sf := &StreamFactory{
@@ -197,7 +197,7 @@ func NewTier1(
 		enforceCompression:      enforceCompression,
 		activeRequestsSoftLimit: activeRequestsSoftLimit,
 		activeRequestsHardLimit: activeRequestsHardLimit,
-		globalRequestPool:       globalRequestPool,
+		sessionPool:             sessionPool,
 	}
 	s.OnTerminating(func(_ error) {
 		s.activeRequests.Wait()
@@ -343,20 +343,20 @@ func (s *Tier1Service) Blocks(
 		}
 	}
 
-	status := s.getOverloadedStatus()
+	stat := s.getOverloadedStatus()
 
 	// Set us as unready if the soft limit would be reached by this request
-	if status.softLimitWouldBeReached() {
+	if stat.softLimitWouldBeReached() {
 		s.logger.Debug("soft limit would be reached by this request, setting app as unready",
-			append(fields, zap.Int("active_request_count", status.activeRequestCount), zap.Int("soft_limit", status.softLimit))...,
+			append(fields, zap.Int("active_request_count", stat.activeRequestCount), zap.Int("soft_limit", stat.softLimit))...,
 		)
 		s.appSetIsReadyState(false)
 	}
 
 	// Refuse the request if the hard limit is currently reached by this instance
-	if status.hardLimitReached() {
+	if stat.hardLimitReached() {
 		err := connect.NewError(connect.CodeUnavailable, fmt.Errorf("service under heavy load, please try connecting again"))
-		fields = append(fields, zap.Error(err), zap.Int("active_request_count", status.activeRequestCount), zap.Int("hard_limit", status.hardLimit))
+		fields = append(fields, zap.Error(err), zap.Int("active_request_count", stat.activeRequestCount), zap.Int("hard_limit", stat.hardLimit))
 		logger.Info("refusing Substreams Blocks request", fields...)
 		return err
 	}
@@ -439,8 +439,10 @@ func (s *Tier1Service) Blocks(
 		}
 	}()
 
+	runningContext = reqctx.WithCancelFunc(runningContext, cancelRunning) // we pass this down so that the 'workerPool/requestPool' can cancel the running context
+
 	respFunc := tier1ResponseHandler(respContext, &mut, logger, stream, request.NoopMode, reqStats, req.Msg.DevOutputModules)
-	err = s.blocks(runningContext, request, req.Header(), execGraph, respFunc, reqStats, fields)
+	err = s.blocks(runningContext, cancelRunning, request, req.Header(), execGraph, respFunc, reqStats, fields)
 
 	if connectError := toConnectError(runningContext, err); connectError != nil {
 		switch connect.CodeOf(connectError) {
@@ -450,6 +452,8 @@ func (s *Tier1Service) Blocks(
 			logger.Debug("invalid argument on request", zap.Error(connectError))
 		case connect.CodeCanceled:
 			logger.Debug("Blocks request canceled by user", zap.Error(connectError))
+		case connect.CodeResourceExhausted:
+			logger.Debug("Blocks request failed with ResourceExhausted", zap.Error(connectError))
 		default:
 			logger.Warn("Blocks request completed with error", zap.Error(connectError))
 		}
@@ -502,7 +506,7 @@ func (s *Tier1Service) writeLastUsed(ctx context.Context, execGraph *exec.Graph,
 
 var IsValidCacheTag = regexp.MustCompile(`^[a-zA-Z0-9_-]+$`).MatchString
 
-func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Request, header http.Header, execGraph *exec.Graph, respFunc substreams.ResponseFunc, reqStats *metrics.Stats, logFields []zap.Field) (err error) {
+func (s *Tier1Service) blocks(ctx context.Context, cancelRunning context.CancelCauseFunc, request *pbsubstreamsrpc.Request, header http.Header, execGraph *exec.Graph, respFunc substreams.ResponseFunc, reqStats *metrics.Stats, logFields []zap.Field) (err error) {
 	chainFirstStreamableBlock := bstream.GetProtocolFirstStreamableBlock
 	if request.StartBlockNum > 0 && request.StartBlockNum < int64(chainFirstStreamableBlock) {
 		return bsstream.NewErrInvalidArg("invalid start block %d, must be >= %d (the first streamable block of the chain)", request.StartBlockNum, chainFirstStreamableBlock)
@@ -525,7 +529,7 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 
 	if request.StopBlockNum != 0 {
 		if requestDetails.ResolvedStartBlockNum == request.StopBlockNum {
-			err := bsstream.NewErrInvalidArg("start block and stop block are the same")
+			err := bsstream.NewErrInvalidArg("start block and stop block are the same: %d and %d", requestDetails.ResolvedStartBlockNum, request.StopBlockNum)
 			logger.Info("refusing Substreams Blocks request", append(logFields, zap.Error(err))...)
 			return err
 		}
@@ -697,25 +701,39 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 		return fmt.Errorf("error building request plan: %w", err)
 	}
 
-	if s.globalRequestPool != nil {
-		userID := dauth.FromContext(ctx).UserID()
-		apiKeyID := dauth.FromContext(ctx).APIKeyID()
-		r := s.globalRequestPool.BorrowRequest(ctx, userID, apiKeyID, tracing.GetTraceID(ctx).String())
-		if r.status == pbworker.BorrowWorkerResponse_resource_exhausted {
-			msg := strings.Builder{}
-			msg.WriteString("Request quota exceeded.\n")
-			msg.WriteString(fmt.Sprintf("Your allowed %d concurrent requests.\n", r.state.MaxWorkers))
-			msg.WriteString(fmt.Sprintf("Each request has a minimal life time of %s\n", r.minimalWorkerLifeDuration.String()))
-			return status.Errorf(codes.ResourceExhausted, "%s", msg.String())
+	if s.sessionPool != nil {
+		auth := dauth.FromContext(ctx)
+		userID := auth.UserID()
+		apiKeyID := auth.APIKeyID()
+		traceID := tracing.GetTraceID(ctx).String()
+		service := "t1r"
+
+		sessionID, err := s.sessionPool.Get(ctx, service, userID, apiKeyID, traceID, func(err error) {
+			if cancelRunning != nil { // in tests, this might be nil
+				cancelRunning(err)
+			}
+		})
+
+		if err != nil {
+			switch {
+			case errors.Is(err, dsession.ErrConcurrentStreamLimitExceeded),
+				errors.Is(err, dsession.ErrPermissionDenied),
+				errors.Is(err, dsession.ErrQuotaExceeded):
+				s.logger.Info("session denied to user", zap.String("user_id", userID), zap.String("api_key_id", apiKeyID), zap.String("trace_id", traceID), zap.Error(err))
+			default:
+				s.logger.Error("failed to acquire session", zap.Error(err), zap.String("service", service), zap.String("user_id", userID), zap.String("api_key_id", apiKeyID), zap.String("trace_id", traceID))
+			}
+			return err
 		}
 
+		s.logger.Debug("acquired session", zap.String("session_id", sessionID))
+
+		// Pass sessionKey through context for WorkerPool
+		ctx = reqctx.WithSessionKey(ctx, sessionID)
+
 		defer func() {
-			zlog.Info("returning request", zap.Bool("keep", false), zap.String("key", r.key))
-			if s.IsTerminating() {
-				s.logger.Info("returning request without minimal life time. Server is shutting down", zap.String("key", r.key))
-				r.minimalWorkerLifeDuration = 0
-			}
-			s.globalRequestPool.ReturnRequest(r)
+			s.logger.Debug("releasing session", zap.String("session_id", sessionID))
+			s.sessionPool.Release(sessionID)
 		}()
 	}
 
@@ -770,7 +788,7 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 		zap.String("cursor", cursor),
 	)
 
-	var streamHandler bstream.Handler
+	var wrappedPipe bstream.Handler
 	if requestDetails.ProductionMode {
 		liveBackFiller := NewLiveBackFiller(ctx, pipe, logger, execGraph.OutputModuleStageIndex(), segmentSize, requestDetails.LinearHandoffBlockNum, s.runtimeConfig.ClientFactory, RequestBackProcessing)
 
@@ -781,14 +799,14 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 		}
 
 		go liveBackFiller.Start(ctx)
-		streamHandler = liveBackFiller
+		wrappedPipe = liveBackFiller
 	} else {
-		streamHandler = pipe
+		wrappedPipe = pipe
 	}
 
 	blockStream, err := s.streamFactoryFunc(
 		ctx,
-		streamHandler,
+		wrappedPipe,
 		int64(requestDetails.LinearHandoffBlockNum),
 		request.StopBlockNum,
 		cursor,
@@ -803,7 +821,37 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 	}
 
 	ctx, span := reqctx.WithSpan(ctx, "substreams/tier1/pipeline/blocks_stream")
-	streamErr = blockStream.Run(ctx)
+	for {
+		streamErr = blockStream.Run(ctx)
+		if errors.Is(streamErr, hub.ErrSubscriptionChannelFull) {
+			cur := pipe.LastCursor()
+			if cur == nil {
+				logger.Warn("subscription channel at max capacity, but no cursor was found to reconnect")
+				break
+			}
+
+			logger.Warn("subscription channel at max capacity, creating new stream", zap.String("last_block", cur.Block.String()))
+			blockStream, err = s.streamFactoryFunc(
+				ctx,
+				wrappedPipe,
+				int64(requestDetails.LinearHandoffBlockNum),
+				request.StopBlockNum,
+				cur.ToOpaque(),
+				request.FinalBlocksOnly,
+				false,
+				logger.Named("stream"),
+				bsstream.WithLiveSourceHandlerMiddleware(metering.LiveSourceMiddlewareHandlerFactory(ctx)),
+				bsstream.WithFileSourceHandlerMiddleware(metering.FileSourceMiddlewareHandlerFactory(ctx)),
+			)
+			if err != nil {
+				streamErr = fmt.Errorf("error getting stream: %w", err)
+				break
+			}
+			continue
+		}
+		break
+	}
+
 	span.EndWithErr(&streamErr)
 
 	return pipe.OnStreamTerminated(ctx, streamErr)
@@ -868,10 +916,12 @@ func tier1ResponseHandler(ctx context.Context, mut *sync.Mutex, logger *zap.Logg
 			}
 		}
 
+		begin := time.Now()
 		if err := streamSrv.Send(resp); err != nil {
 			logger.Info("unable to send block probably due to client disconnecting", zap.String("user_id", userID), zap.String("api_key_id", apiKeyID), zap.Error(err))
 			return connect.NewError(connect.CodeUnavailable, err)
 		}
+		stats.RecordReadTime(begin)
 
 		if isData {
 			stats.RecordDataSent()
@@ -1006,34 +1056,39 @@ func toConnectError(ctx context.Context, err error) error {
 		return nil
 	}
 
+	if errors.Is(err, context.Canceled) {
+		if contextCause := context.Cause(ctx); contextCause != nil {
+			err = contextCause // unwrap errors in canceled contexts
+		} else {
+			return connect.NewError(connect.CodeCanceled, err)
+		}
+	}
+
+	if err, ok := dsession.ToConnectError(err); ok {
+		return err
+	}
+	// special case for context canceled when shutting down
+	if err == errShuttingDown {
+		return connect.NewError(connect.CodeUnavailable, err)
+	}
+
 	// GRPC to connect error
 	if grpcError := dgrpc.AsGRPCError(err); grpcError != nil {
 		switch grpcError.Code() {
 		case codes.Canceled:
-			return connect.NewError(connect.CodeCanceled, grpcError.Err())
+			return connect.NewError(connect.CodeCanceled, errors.New(grpcError.Message()))
 		case codes.Unavailable:
-			return connect.NewError(connect.CodeUnavailable, grpcError.Err())
+			return connect.NewError(connect.CodeUnavailable, errors.New(grpcError.Message()))
 		case codes.InvalidArgument:
-			return connect.NewError(connect.CodeInvalidArgument, grpcError.Err())
+			return connect.NewError(connect.CodeInvalidArgument, errors.New(grpcError.Message()))
 		case codes.DeadlineExceeded:
 			return connect.NewError(connect.CodeDeadlineExceeded, err)
 		case codes.ResourceExhausted:
-			return connect.NewError(connect.CodeResourceExhausted, grpcError.Err())
+			return connect.NewError(connect.CodeResourceExhausted, errors.New(grpcError.Message()))
 		case codes.Unknown:
-			return connect.NewError(connect.CodeUnknown, grpcError.Err())
+			return connect.NewError(connect.CodeUnknown, errors.New(grpcError.Message()))
 		}
 		return grpcError.Err()
-	}
-
-	// special case for context canceled when shutting down
-	if errors.Is(err, context.Canceled) {
-		if context.Cause(ctx) != nil {
-			err = context.Cause(ctx)
-			if err == errShuttingDown {
-				return connect.NewError(connect.CodeUnavailable, err)
-			}
-		}
-		return connect.NewError(connect.CodeCanceled, err)
 	}
 
 	// special case for "QuickSave" on shutdown